+++ tests/Fractions.Tests/FractionSpecs/Sqrt/Method_Sqrt.cs
@@ -1,7 +1,9 @@
 ﻿using System;
+using System.Collections.Generic;
 using FluentAssertions;
 using NUnit.Framework;
 using Tests.Fractions;
+using Fractions.Extensions;
 
 namespace Fractions.Tests.FractionSpecs.Sqrt {
 
@@ -38,7 +40,7 @@ namespace Fractions.Tests.FractionSpecs.Sqrt {
             var expected = Math.Sqrt(value);
             var actual = Fraction.FromDouble(value).Sqrt();
             actual.ToDouble().Should().Be(expected);
-        }```
+        }
 
 
     }
